Added error checking for hashcheck script, shorthand for workflow

+++ backend/hashcheck.sh
@@ -1,21 +1,26 @@
 #!/bin/bash
 check() {
     #{ aws ecr get-login-password --region ap-southeast-1 | docker login --username AWS --password-stdin $ECR_REGISTRY; }
-    docker pull $ECR_REGISTRY/$1
-    repo=$(docker images $ECR_REGISTRY/$1 --no-trunc --format "{{.ID}}")
-    repo=${repo##sha256:}
-    image=$(docker images $1 --no-trunc --format "{{.ID}}")
-    image=${image##sha256:}
-    docker rmi $ECR_REGISTRY/$1 2>&1 > /dev/null
+    docker pull ${ECR_REGISTRY:?Run \"export ECR_REGISTRY variable\"}/$1 2>&1 > /dev/null\
+        || { echo Error pulling image $ECR_REGISTRY/$1; exit 0; }
 
-    status=0
+    image=$(docker images $1 --no-trunc --format "{{.ID}}")
+    : ${image:?No such image $1}
+    image=${image##sha256:}
+
+    repo=$(docker images $ECR_REGISTRY/$1 --no-trunc --format "{{.ID}}")
+    : ${repo:?Error using image $ECR_REGISTRY/$1}
+    repo=${repo##sha256:}
+    docker rmi $ECR_REGISTRY/$1 2>&1 > /dev/null
+    echo "Repo:$repo Image:$image"
+
     if [ $repo = $image ] ; then
         echo Hashes are identical
+        return 0
     else
         echo Hashes are different
-        status=1;
+        return 1
     fi
-    return $status
 }
 
 if [ $# = 0 ]; then
